Front page carousel done and some theming
Wow! It took me a while and lots of Googling, but I finally sorted out the carousel for the front page. After some hunting on SO, I decided to just hardcode the images in. GraphQL was not cooperating with me with querying thumbnails on randomMeal, and the solution I found on SO to that was very complicated.