Bug 1 fixed!
After a lot of searching on Stack Overflow, I ran across one post where the poster had a separate term for the current value in the form, and they only updated the search term (thereby triggering the search) when the user submitted the form. That seems to have been the issue - I had my form tied directly to the search value, and Apollo was probably performing a search every time it saw the term change!